Команды статистики
- Добавлены новые команды (!мойпрофиль, !топпользователи, !статистикабота)
- Улучшен формат вывода статистики
- Исправлена архитектурная проблема и ошибка написания (отдельная благодарность таланту)

+++ PoproshaykaBot.WinForms/Models/UserStatistics.cs
@@ -6,7 +6,7 @@ public class UserStatistics
 
     public string Name { get; set; } = string.Empty;
 
-    public int MessageCount { get; set; }
+    public ulong MessageCount { get; set; }
 
     public DateTime FirstSeen { get; set; } = DateTime.UtcNow;
